requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23

# Google Calendar API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.25.2

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
